requirements.txt
cloudpickle==3.0.0
Farama-Notifications==0.0.4
gymnasium==0.29.1
numpy==2.1.1
pyglet==1.5.29
pygame==2.6.0 # might not be required in future development
stable-retro==0.9.2
typing_extensions==4.12.2
